fix method for collecting args for constructors
separate functionalities of inheritanceInliner in different files for better readability
