+++ src/passes/index.ts
@@ -16,7 +16,7 @@ export * from './generateGetters';
 export * from './identifierMangler';
 export * from './ifFunctionaliser';
 export * from './implicitConversionToExplicit';
-export * from './inheritanceInliner';
+export * from './inheritanceInliner/inheritanceInliner';
 export * from './intBoundCalculator';
 export * from './literalExpressionEvaluator/literalExpressionEvaluator';
 export * from './loopFunctionaliser';
